rename test generator functions
